Setup project to use Maven.
We've decided to use Maven since Lanterna uses Maven and is in Maven
Central. This means that we don't have to package Lanterna source or a
Lanterna jar into our project. It also allows for us to more clearly
define what is in the engine and what is in the game.