Created a Lanterna frontend to serve as a framework for our roguelike.
Lanterna provides methods for drawing to a screen, but little else. As a
result, it is necessary to wrap it in something generic that can be used
for developing a game. LanternaFrontend and Application serve this
goals.

LanternaFrontend is meant to be somewhat more generic than Application,
and just assumes that some type of drawing to the console with key
presses is happening. Application makes several more assumptions about
how the application itself behaves and functions.

This commit includes an interactive test for checking the behaviour of
the several new files. I plan to add these tests to a new suite
called "interactive" at some point, so that they will not run
automatically and must be run under a separate Maven profile. The test
will also need to be modified so that it can actually end, instead of
requiring ^C.

GameAction is meant to represent an action that occurs inside the game.
Please read its Javadoc and the link provided there for more information
on how it is intended to be used.