Layed out beginning framework for controlling entities.
All controllable entities (either by the player or by AI) will be
controlled through an interface that allows for messages to be passed to
the entity either as queries or as commands. This allows for the
separation of some logic from data and will also allow for listeners for
certain events.

The primary reason for the separation is that it will allow for more
flexibility in how entities are controlled. Because we use an interface
to represent the object that controls entities, the destination could be
local or on the other side of a network. We will also be able to do
things like switch what entity the keyboard controls, distribute the
same commands across several managers, pass managers to entities that
are responsible for effects (like poison), etc., if we so desire.