Created Section which allows creating offsets onto the Screen.
Section extends ScreenWriter, and provides the ability to define offsets
that will be used when drawString() is called. This allows you to define
subsections of the Screen, which can be passed to the doDraw() method of
a Drawable. By doing things this way, Drawables become very portable,
and can assume that they can draw entirely onto one single space. All of
the logic for drawing can also be included inside the Drawable, so that
it need not be copied or held externally. The encapsulating Layer (or
anything else) just needs to tell it where to draw.