Monsters are removed from the EntityManager when they die.
When Monsters die, they are now removed from the EntityManager. Before,
they would not get removed, which meant that they still responded to
keyboard input and anything else that would fetch them from the
EntityManager. This resulted in problems like Monsters dying and
disappearing, but then reappearing as soon as they could set themselves
to a Tile on the Level. (See T35).