EntityManger no longer throws an Exception when category doesn't exist.
When EntityManger.getEntity() was called with a non-existent category, a
NullPointException would be thrown while it was trying to copy the list.
Now it will create an empty list for that category if one doesn't
currently exist and put it into the HashMap. It then makes a copy of
whatever list exists and returns it. It is important to copy the list
that is going to be returned, to avoid a ConcurrentModificationException
if an entity is removed from the EntityManager as a result of processing
the returned map.