Added two TileTypes to indicate up and down stairs.
Stairs will at some point need to be randomly generated into a level. A
reference to them will also need to be accesible through some getter so
that the player can be placed there when initially entering into a
level. EntityActionManager will also need some way of instructing the
MainCharacter to go up and down stairs. The MainCharacter can check if
s/he is on stairs by asking the `location` its TileType.