There is now a MainCharacter that is capable of fighting Combatables.
There is now a MainCharacter who is in the categories "keyboard" and
"main". Eventually we may want to extend the configuration options to
allow us to provide information on the MainCharacter externally, along
with the categories for the MainCharacter and Monsters. If we do things
like this, then we could always specify the MainCharacter as a monster
who is in the category "keyboard" while other Monsters are in the
category "ai" or something similar.

The MainCharacter is currently generated in ProgressiveMonsterGenerator
and placed at random into a room. Eventually we will want to place the
MainCharacter in the same location as the ascending staircase, for
consistency if s/he goes down a level. On the first level, ascending
this staircase should trigger an end-game, since this would equate to
exiting the entire dungeon.

See task T36.