Created Action for chasing (and therefore attacking) main character.
So that we'll have a basis for attacking the player so that we can begin
playtesting, there is now a simple Action that will automatically chase
the player whenever he attacks. Since (for now) moving onto a Tile that
contains another Entity triggers an attack, this takes care of fighting.