Interim: skeleton TurnManager, everything list in EntityManager
Added a skeleton TurnManager class, and also added an everything
list to EntityMangager. The everything list will be used to reflect
what EntityActionManagers  actually exist. For example,
TurnManager can grab the list copies of EntityActionManagers by
querying by categories, but before it grants a turn to an
EntityActionManager it can check to see if the really exists
by checking for presence in the everything list.