Poitons and Useable Items
There are now the ability to collect potions and use them. Pressing 'q' opens the potion layer, 'Q' now quits. After opening ht potion layer ,press an alpha key to use the item in that slot. Potions store an action, and this action is applied to the player when the player quaffs the potion. Right now, all potions are +5 HP. e