Wieldable Weapons
There are now wieldable weapons. Each weapon has a stats object that is added to the player's base stats when equipped, and subtracted when unequipped. Similar extensions will soon be made for armor and rings. When you equip an item, it is stored in the player's equipment map and removed from inventory. It is put back in the inventory when unequipped.

'w' brings up the wield menu.

Right now, random generator only generates swords. Adaptive difficulty can be coded so that the generator generates progressively better swords at each level. Weapons are represented by the ')' character.