Exploration
The screen now starts with all tiles hidden and the players current room shown. Each time the player moves, the level checks if the player is in a room, if so, it will reveak the entire room the player is in. Else, the player is shown a view of a radius around phe. This can be extended with partial raycasting and stuff so that rooms are not immediately shown. Now, the player actually needst o explore the level, trying to find items, stairs etc.