Multiple Levels 1
- Game now has a gotoLevel(depth) method that moves the player to the current level.
- The game maintains a list of levels that exist.
- The player is removed from the level and unregistered when changing levels, allowing simulation.
- BSPGenerator now creates the up/down stairs.
- Player is placed at the upStairs when going down, and the down stairs when coming up.

As of now, there is no ability to change levels based on stairs, this will be next. For now, pressing 'N' brings the player to the next level. You can see the player and its inventory is now kept. The items will get better as you go down.

Mechanics that need to be altered is the win condition.