Added Artist abstraction, so that logic related to drawing is cleaner.
Certain things (like Level) should contain the logic for drawing
themselves, instead of keeping it in something like DefaultMainLayer. By
defining an interface called Artist, we can define objects which are
responsible for drawing themselves. We can then provide a Section so
that they are limited to an area which we define for drawing.