Created a template for reading in context keybindings.
In order to make keybindings more general and flexible, I have decided
that they should be read in from a configuration file. Similar to the
configuration for weapons and monsters, they will be JSON files located
in the configuration directory.

The JSON will be an array of objects with two fields: the name of the
context, and an object that represents the mapping of keybindings to
the command alias.

To keep the flexibility offered by GameActions while making it easy to
read/understand what numbers correspond to, I will introduce aliases -
mappings of names to numbers that can be used in the configuration files
and will be handled when the files are read in.